requirements.txt
Django>=3.2,<4.0
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
drf_yasg
